requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Image Processing
pillow==10.1.0
numpy==1.25.2

# HTTP Requests
requests==2.31.0
httpx==0.25.1

# AI/ML
openai==1.3.7

# Web Automation (Optional)
playwright==1.40.0

# Environment Variables
python-dotenv==1.0.0

# Development
uvicorn[standard]>=0.24.0